Portoflio changes.

+++ project/src/main/java/com/narain/portfoliotracker/controller/PortfolioController.java
@@ -1,19 +1,25 @@
 package com.narain.portfoliotracker.controller;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.UUID;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.access.AccessDeniedException; 
 
 import com.narain.portfoliotracker.dto.PorfolioValueWrapper;
 import com.narain.portfoliotracker.model.Asset;
@@ -44,14 +50,26 @@ public class PortfolioController {
         return portfolioService.getPortfolioLiveValue(portfolio);
     }
 
-    @PostMapping("/added-asset")
-    public ResponseEntity<String> addAsset(@RequestBody AssetRequest request) {
-        portfolioService.addAssetToPortfolio(request.getPortfolio(), request.getAsset());
-        return ResponseEntity.ok("Asset added successfully to portfolio.");
-    }
+   @PostMapping("/add-asset/{portfolioId}")
+    public ResponseEntity<String> addAsset(@PathVariable UUID portfolioId, @RequestBody Asset asset, Authentication authentication) throws AccessDeniedException {
+        String username = authentication.getName();
 
+        try {
+            portfolioService.addAssetToPortfolio(username, portfolioId, asset);
+            return ResponseEntity.ok("Asset added successfully to portfolio.");
+        } catch (AccessDeniedException e) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Unauthorized to modify this portfolio.");
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An error occurred while adding the asset: " + e.getMessage());
+        }
+    }
     @DeleteMapping("/removed-asset/{ticker}")
-    public ResponseEntity<String> removeAsset(@RequestBody Portfolio portfolio, @PathVariable String ticker) {
+    public ResponseEntity<String> removeAsset(@RequestBody Portfolio portfolio, UUID portfolioId, @PathVariable String ticker, Authentication authentication) {
+        String username = authentication.getName();
+        if (!portfolioService.userOwnsPortfolio(username, portfolioId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Unauthorised access to this resource.");
+        }
+
         boolean removed = portfolioService.removeAssetFromPortfolio(portfolio, ticker);
         if (removed) {
             return ResponseEntity.ok("Asset removed successfully from portfolio.");
@@ -122,4 +140,11 @@ public class PortfolioController {
         return ResponseEntity.status(HttpStatus.CREATED).body(savedPortfolio);
     }
 
+    @GetMapping("/my-portfolios")
+    public ResponseEntity<List<Portfolio>> getMyPortfolios(Authentication authentication) {
+        String username = authentication.getName();
+        List<Portfolio> portfolios = portfolioService.getPortfoliosByUsername(username);
+
+        return ResponseEntity.ok(portfolios);
+    } 
 }
